Rust block_num u16

+++ rust/src/block.rs
@@ -1,4 +1,4 @@
-pub fn block(block_num: u32) -> &'static [u8] {
+pub fn block(block_num: u16) -> &'static [u8] {
 	return match block_num {
 		0x000 => include_bytes!("data/000.bin"),
 		0x001 => include_bytes!("data/001.bin"),
